requirements.txt
opencv-python==4.5.5.62
torch==1.10.0
transformers==4.15.0
Pillow==8.4.0
numpy==1.21.4
